requirements.txt
imageio
matplotlib
numpy
pillow
scipy
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tqdm
